+++ src/webview/js/components/uri.js
@@ -24,8 +24,11 @@ AFRAME.registerComponent('uri', {
 
         //as init() runs asynchronously, it might run after the panel was loaded
         //config info needs to be reloaded
-        this.attrValue.configMethod[0](this.el)
-
+        if(this.attrValue.configMethod[0])
+        {
+            this.attrValue.configMethod[0](this.el)
+        }
+        
         // this.attributes        = {}
     },