+++ src/webview/js/ui/button.js
@@ -1,6 +1,10 @@
 AFRAME.registerComponent('button', {
     schema: {
         //menu: {}
+        value: { type: "string", default: "" },
+        width: { type: "number", default: 1 },
+        wrapcount: { type: "number" },
+        enabled: { type: "boolean", default: true }
     },
     init: function () {
 
@@ -12,19 +16,26 @@ AFRAME.registerComponent('button', {
         let button = document.createElement('a-box')    
         button.setAttribute('opacity', '.3')
         button.setAttribute('depth', '.1')
-        button.setAttribute('width', '1')
+        // button.setAttribute('width', '1')
+        button.setAttribute('width', this.data.width)
         button.setAttribute('height', '.3')
         // button.setAttribute('animation__scale',         {property: 'opacity', dur: 0, to: '.6', startEvents: 'mouseenter'});
         // button.setAttribute('animation__scale_reverse', {property: 'opacity', dur: '0', to: '.4',   startEvents: 'mouseleave'});
         button.setAttribute('animation__scale',         {property: 'opacity', dur: 0, to: '.6', startEvents: 'mousedown'});
         button.setAttribute('animation__scale_reverse', {property: 'opacity', dur: 0, to: '.4', startEvents: 'mouseup'});
-        button.classList.add('interactive')
+        
+        if(this.data.enabled){
+            button.classList.add('interactive')
+        }
+        else{
+            button.setAttribute('visible', false)
+        }
         button.object3D.position.setZ(button.object3D.position.z+0.05)
     
 
         //set label to menu button
         // appendLabel(groupButton, menu.name)
-        let label = appendLabel(button, "done")
+        let label = appendLabel(button, this.data.value, this.data.wrapcount)
         // label.object3D.position.set(0,0,5)
         label.setAttribute('position', "0 0 .049")
 
@@ -40,7 +51,27 @@ AFRAME.registerComponent('button', {
 
         this.el.appendChild(button)
     },
-    update: function () {
+    update: function (oldData) {
+
+        //at creation time there is no data, so we ignore the invocation
+        if(Object.keys(oldData).length === 0){
+            return
+        }
+
+        //if 'enabled' did change        
+        if(this.data.enabled != oldData.enabled){
+
+            //enable interaction ON/OFF
+            if(this.data.enabled){
+                this.el.firstChild.classList.add('interactive')
+            }
+            else{
+                this.el.firstChild.classList.remove('interactive')
+            }
+
+            //switch ON/OFF button's visibility
+            this.el.firstChild.setAttribute('visible', this.data.enabled)
+        }
     },
     tick: function () {},
     remove: function () {},
@@ -54,5 +85,9 @@ AFRAME.registerComponent('button', {
     },
     mappings: {
         // menu: "dropdown.menu"
+        value: "button.value",
+        width: "button.width",
+        wrapcount: "button.wrapcount",
+        enabled: "button.enabled"
     }
   });