+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Change Log
 
+## 0.10.0
+
+- New AtlasMap endpoint to create data mappings. \
+  The endpoint integrates with the VSCode AtlasMap extension to open/create ADM files. \
+  To properly integrate with AtlasMap ensure you have installed:
+  - AtlasMap Data Transformation editor by Red Hat (minimum version 0.1.2)
+- included in menu a 'help' category with access to 'how-to' documentation
+- included in configuration panels access to 'how-to' documentation
+
+
 ## 0.9.2
 
 - **IMPORTANT**: install this version for users on VSCode 1.62.x onwards.